requirement.txt
matplotlib==3.5.1
numpy==1.22.3
Pillow==9.1.0
streamlit==1.8.1
tensorflow-cpu==2.8.0
tensorflow_hub==0.12.0